+++ models/user_role.py
@@ -1,0 +1,7 @@
+import enum
+
+
+class UserRole(enum.Enum):
+    admin = 'admin'
+    user = 'user'
+    guest = 'guest'
